+++ test/_env/mocks/redis.js
@@ -30,11 +30,14 @@ const client = {
   xAdd: jest.fn(() => {
     return Promise.resolve();
   }),
+  off: jest.fn(() => {}),
   subscribe: jest.fn((channel, cb) => {
     onSubscribe.push(cb);
   }),
   sSubscribe: jest.fn(() => {}),
   pSubscribe: jest.fn(() => {}),
+  unsubscribe: jest.fn(() => {}),
+  pUnsubscribe: jest.fn(() => {}),
   publish: jest.fn((channel, message) => {
     for (const on of onMessage) {
       on(channel, message);