+++ test/_env/mocks/redisStreamAdapter.js
@@ -4,9 +4,7 @@ const adapter = {
   addAll: jest.fn(),
   delAll: jest.fn(),
   del: jest.fn(),
-  broadcast: jest.fn(() => {
-    debugger;
-  }),
+  broadcast: jest.fn(() => {}),
   persistSession: jest.fn(),
   socketRooms: jest.fn(() => {
     return new Set();